+++ CHANGELOG.md
@@ -15,6 +15,7 @@ feat(networking): add routeTableAssociations and routeTablePropagations to trans
 - fix(config): add policy replacement support for custom Config rule
 - fix(securityhub): normalize control id prefix matching for controlsToDisable
 - fix(validator): validate account name and email length constraints before DynamoDB state persistence
+- fix(controltower): update logic for kms key on update landing zone operations.
 
 ### Changed
 chore(logging): update elb access logs bucket for service principal